refactor: detail repository 분리
