+++ growthmate-api/src/main/java/com/growup/growthmate/company/dto/detail/CompanyDetailResponse.java
@@ -4,12 +4,12 @@ import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public record CompanyDetailResponse(
-                String name,
-                String imageUrl,
-                String businessType,
-                LocalDateTime establishmentDate,
-                String address,
-                BigInteger employeeNumber,
-                BigInteger sales
+        String name,
+        String imageUrl,
+        String businessType,
+        LocalDateTime establishmentDate,
+        String address,
+        BigInteger employeeNumber,
+        BigInteger sales
 ) {
 }
