+++ growthmate-api/src/main/java/com/growup/growthmate/company/repository/analysis/CompanyAnalysisRepository.java
@@ -1,4 +1,4 @@
-package com.growup.growthmate.company.repository;
+package com.growup.growthmate.company.repository.analysis;
 
 import com.growup.growthmate.company.domain.Company;
 import com.growup.growthmate.company.domain.CompanyAnalysis;
@@ -6,8 +6,6 @@ import com.growup.growthmate.company.dto.analysis.CompanyAnalysisRequest;
 import com.growup.growthmate.company.dto.detail.CompanyDetailRequest;
 
 public interface CompanyAnalysisRepository {
-
-    Company findCompanyDetail(CompanyDetailRequest request);
 
     CompanyAnalysis findCompanyAnalysis(CompanyAnalysisRequest request);
